Add stream support on bokeh TablePlot
